Added installation instructions
There are number of oauth solutions for node.js and it is
helpful to provide the official installation instructions
in the Readme file.  Installation instructions can be a little
confusing for this module because the github repository name
is node-oauth and there is a different  node.js oauth
solution that isntalls via npm with the command node-oauth.
Since this solution installs with oauth official
instructions may be helpful to some developers.