Handle multiple params of the same name properly
In case we try to call consumer.get('URL?param=1&param=2'), the library
generates incorrect URL. That is because it internally parses the URL so
that the params are actually called 'param[0]' and 'param[1]'.

This patch fixed the param names so that the resulting URL is actually
including 'param=1' and 'param=2'.